refactored cleaned optical raycaster
